refactor: rename state.ts -> workflow-state.ts
